api/requirements.txt
flask==2.2.2
flask_cors
mysql-connector-python==8.0.33
